// src/app/Models/Content/ContentForEditing.ts
export class ContentForEditing {
    id: number;
    author: string;
    title: string;
    typeIds: number[];
    shortDescription: string;
    articleAttachmentPaths: string[];
    previewAttachmentsBase64: string[];
    previewAttachmentPaths: string[];
    text: string;
    editors: string[];
}